src/functions/OpenMeteo-func/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.21.1
pytz==2024.1
openmeteo-requests
retry-requests
requests-cache
pandas